requirements.txt
docx2txt==0.8
faiss-cpu
groq==0.9.0
numpy==1.26.4
openai==1.37.1
pdfplumber==0.11.2
python-dotenv==1.0.1
streamlit==1.37.0